clean up host package generator, fix compile errors after rebase
